+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 0.16
 ------------
 - datalink is now a real type, not a domain
+- use varchar's typmod in/out functions
 
 Version 0.15
 ------------